requirements.txt
pandas>=2.2.2
python-dotenv>=1.0.1
requests>=2.32.3
beautifulsoup4>=4.12.3
PyYAML>=6.0.1
schedule>=1.2.2
streamlit>=1.36.0
plotly>=5.23.0
playwright>=1.46.0
tenacity>=8.3.0
tqdm>=4.66.4 